Added ClassTypeNodeFactory
This is an implementation of an AbstractNodeFactory based on Reflection.
It will check if a class passed to it has both the right superclass and
constructor. (This makes it easy to extend the available NodeFactories in
the Future, since we just need to implement the Node and then pass the class
to the ClassTypeNodeFactory).