Setup Nodes and NodeFactories
One of the goals of refactoring my original code is to untie the dependency
between the AhoCorasick String Matcher and Ukkonens Algorithm. There is
a clear dependency from UA on the ACSM, but not the other way round.
Since we might want to extend the ACSM in the future by adding new node
types, we want to make the Nodes less concrete. Therefore, we add a Node
Factory interface.

Nodes can only be created via Factories using the AbstractACNodeFactory
interface - This allows us to dynamically switch the type of node used
in the AhoCorasickStringMatcher machine without resorting to Reflection.
All nodes used must extend ACTrieNode, which provides basic functionalities,
so that we can execute the original algorithm.

Note on Design:
ACTrieNode is very stable.
AbstractACNodeFactory is very stable.
LanguageParameters is very stable.
(Defines Mapping between Characters of an Alphabet and Integers)
AhoCorasickTrie is now created by an AhoCorasickTrieFactory